Refactored cores/esp32/FunctionalInterrupt to not support illegal use of non-IRAM ISRs
via C++ std::bind, std::function, or lambdas.
It is now a follow up to the introduction of Schedule.(h|cpp) into core, which supports
safe scheduling of non-IRAM ISRs as they are being scheduled synchronously to loop() in the
same task.
