+++ cores/esp32/FunctionalInterrupt.h
@@ -1,5 +1,5 @@
-#ifndef FUNCTIONALINTERRUPTS_H
-#define FUNCTIONALINTERRUPTS_H
+#ifndef CORE_FUNCTIONALINTERRUPT_H
+#define CORE_FUNCTIONALINTERRUPT_H
 
 #include <functional>
 
@@ -13,25 +13,17 @@ struct InterruptInfo
     uint32_t micro = 0;
 };
 
-struct FunctionInfo
-{
-    std::function<void(void)> reqFunction = nullptr;
-    std::function<void(InterruptInfo)> reqScheduledFunction = nullptr;
-};
-
 struct ArgStructure
 {
     ~ArgStructure()
     {
-        delete functionInfo;
         delete interruptInfo;
     }
     InterruptInfo* interruptInfo = nullptr;
-    FunctionInfo* functionInfo = nullptr;
+    std::function<void(InterruptInfo)> scheduledFunction = nullptr;
 };
 
-void attachInterrupt(uint8_t pin, std::function<void(void)> intRoutine, int mode);
 void attachScheduledInterrupt(uint8_t pin, std::function<void(InterruptInfo)> scheduledIntRoutine, int mode);
 void detachFunctionalInterrupt(uint8_t pin);
 
-#endif //INTERRUPTS_H
+#endif //CORE_FUNCTIONALINTERRUPT_H
